devices

+++ models/devices/modbus.go
@@ -42,10 +42,10 @@ func (d DevModBusConfig) Valid() (ok bool, errs []*validation.Error) {
 }
 
 type DevModBusRequest struct {
-	Function string `json:"function"`
-	Address  uint16 `json:"address"`
-	Count    uint16 `json:"count"`
-	Command  []byte `json:"command"`
+	Function string   `json:"function"`
+	Address  uint16   `json:"address"`
+	Count    uint16   `json:"count"`
+	Command  []uint16 `json:"command"`
 }
 
 // params:
